Module paideia::component

This module exports the component trait and components provided by this crate.

Re-exports

pub use block::BlockComponent;

pub use inline::InlineComponent;

Modules

asset

This module exports page assets components.

block

This module exports components that are of type block, as well their kind.

inline

This module exports components that are of type inline, as well their kind.

page

This module exports page component related utilities.

section

This module exports section components-related utilites.

Traits

Component

A component’s base trait. A component is a piece of data or UI logic that will be rendered into UI.

ComponentKind

A component kind is a way of “typing” (i.e. give types) to components at a trait level.