Module paideia::component::block::table

This module exports a table component.

Structs

CaptionedTable

A table with a title/caption at the top.

Cell

A table cell, that can be a header, and can span over multiple columns or rows.

CellAttrs

Attributes of a table cell.

CellComponent

Cell component kind. Components of this kind are usable as cells in a table.

Row

A row in a table. The single unnamed field could be an array, a vec, or anything that iterates by ref yielding cell components.

RowComponent

Row component kind. Components of this kind are usable as rows in a table.

Table

A table. The single unnamed field could be an array, a vec, or anything that iterates by ref yielding row components.