Module paideia::location

This module provides location, paths, Urls.

Structs

Fragment

A fragment of a path, that is, a piece, an element of it.

Id

An ID of a location.

InternalLoc

A location to an internal page, with optional ID. Always absolute (with the root pointing to the root of the encyclopedia).

InternalPath

An internal path, without any ID. Always absolute (with the root pointing to the root of the encyclopedia).

InvalidFragment

Error when an invalid fragment (piece of a path) string is given to be parsed.

InvalidId

Error when an invalid ID string is given to be parsed.

Enums

InvalidInternalLoc

Location

A location of a page, either internal or external.