Trait paideia::render::FullRender

pub trait FullRender: Render<Html> + Render<Markdown> + Render<Text> {
    fn into_dyn<'obj>(self) -> DynFullComponent<'obj, Self::Kind>
    where
        Self: Sized + Send + Sync + 'obj,
    { ... }
    fn into_dyn_unsync<'obj>(self) -> DynFullComponentUnsync<'obj, Self::Kind>
    where
        Self: Sized + 'obj,
    { ... }
}

An auto-trait for components that implement rendering for all supported formats.

Provided Methods

fn into_dyn<'obj>(self) -> DynFullComponent<'obj, Self::Kind>where
Self: Sized + Send + Sync + 'obj,

Converts the fully renderable component into a trait object wrapped by a shared reference. This method is intended for thread-safe components. If your component is not thread-safe or does not care about it, use FullRender::into_dyn_unsync instead.

This method is intended to improve ergonomics of programmers who opt for dynamic dispatch.

fn into_dyn_unsync<'obj>(self) -> DynFullComponentUnsync<'obj, Self::Kind>where
Self: Sized + 'obj,

Converts the fully renderable component into a trait object wrapped by a shared reference. This method is intended for non-thread-safe components, or for components that do not care about thread-safety. If your component is thread-safe and cares about it, use FullRender::into_dyn instead.

This method is intended to improve ergonomics of programmers who opt for dynamic dispatch.