Trait paideia::render::Render

pub trait Render<W>: Componentwhere
    W: Format + ?Sized,{
    fn render(
        &self,
        renderer: &mut Renderer<'_, '_, '_, W>,
        ctx: Context<'_, '_, Self::Kind>
    ) -> Result;
}

General rendering trait. A component implements this trait when it can renders using the given format W.

Required Methods

fn render(
&self,
renderer: &mut Renderer<'_, '_, '_, W>,
ctx: Context<'_, '_, Self::Kind>
) -> Result

Renders the given component using two things:

A renderer backed by a render format;

A rendering context, providing things such page location or section level.

Renderer implements std::fmt::Write, so arbitrary strings can be written to it, HOWEVER, it should be reserved only for writing syntax to the renderer, strings as plain data should be written using their Render::render method.