Trait paideia::render::Scope

pub trait Scope {
    type Format: Format + ?Sized;

    fn enter<F, T>(&self, format: &mut Self::Format, consumer: F) -> T
    where
        F: FnOnce(&mut Self::Format) -> T;
}

A scope of a render format. The scope allows a component to change rendering configuration by entering the scope.

Required Associated Types

type Format: Format + ?Sized

Rendering format associated with this scope.

Required Methods

fn enter<F, T>(&self, format: &mut Self::Format, consumer: F) -> Twhere
F: FnOnce(&mut Self::Format) -> T,

Enters this scope during the call of the given consumer function; exits when this method terminates.